Finishing #9 - file test structure
